Make sure the dispatcher is online when receiving an action
The dispatcher may stop if no action is received for a
given amount of time. If a new action is sent while the
dispatcher is not alive, this action will not execute:
the Intent will be stored in the queue, without being
executed.

This was a bug. We make sure the dispatcher is alive
before handling an action now.